Update NewMetricsHandler to wrap Handler, no namespace, better metrics
From now on we should start prefixing this metric. It would be nice to
have healthcheck_healthy on all projects, so that it kind of becomes
like a up metric, that can be generically alerted on (especially for
liveness, I think). Additionally healthy==1 and unhealthy==0 should make
the metric more intuitive for Prometheus folks.